* Añadido consultas personalizadas con JOIN. Trabajado en los servicios y en los modelos

+++ src/main/java/com/tx/practisesmanagement/model/Location.java
@@ -13,8 +13,8 @@ public class Location {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private Integer id;
-	private Integer latitude;
-	private Integer longitude;
+	private Double latitude;
+	private Double longitude;
 	
 	public Location() {
 		super();
@@ -24,7 +24,7 @@ public class Location {
 		this.id = id;
 	}
 
-	public Location(Integer latitude, Integer longitude) {
+	public Location(Double latitude, Double longitude) {
 		super();
 		this.latitude = latitude;
 		this.longitude = longitude;
@@ -38,19 +38,19 @@ public class Location {
 		this.id = id;
 	}
 
-	public Integer getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Integer latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public Integer getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Integer longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
